Set a match rule for too-long col widths
Similar what I personally have in my own .vimrc. This is kind of hacky I
suppose, but I'd rather have it here than in my own .vimrc. The way this
is now, it only defines a match rule called 'OverLength'. It is up to
the user if they want to use this rule via setting a Highlight on it in
their .vimrc, so it shouldn't be intrusive.